+++ RELEASE_NOTES.md
@@ -1,5 +1,6 @@
 v0.4.0
 ======
+- tag namespace are not mandatory anymore - rename `dockerIdUsername` to `dockerTagNamespace`
 - no more buffering when logging output from Process
 
 v0.3.0